pyproject.toml
[tool.poetry]
name = "aimmo_ojt"
version = "0.1.0"
description = ""
authors = ["homelala"]

[tool.poetry.dependencies]
python = "^3.8"
werkzeug = "=0.16.1"
flask = "^1.1.1"
Flask-Cors = "^3.0.8"
Flask-Classful = "^0.14.2"
mongoengine = "0.20.0"
funcy = "^1.13"
ua-parser = "^0.8.0"
arrow = "^0.15.2"
marshmallow = "^3.2.0"
marshmallow-enum = "^1.5.1"
marshmallow-oneofschema = "^2.0.1"
apispec = "^3.2.0"
flask-apispec = "^0.8.5"
inflection = "^0.3.1"
user-agents = "^2.0"
blinker = "^1.4"
redis = "^3.3.8"
pika = "^1.1.0"
dramatiq = "^1.7.0"
dramatiq_dashboard = "^0.2.2"
watch = "^0.2.7"
sentry-sdk = "^0.12.2"
Flask-Babel = "^0.12.2"
passlib = "^1.7.1"
requests = "^2.22"
flask-mongoengine = "^0.9.5"
pymongo = { version = "=3.12.1", extras = ["srv"] }
boto3 = "^1.10"

webargs = "=5.5.2"
jsondiff = "^1.2"
jsonstreams = "^0.4.2"
zipstream = "^1.1.4"
MarkupSafe = "1.1.1"
PyJWT = "^1.7.1"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
mongomock = "^4.1.2"
flask-swagger-ui = "^4.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

